refactor: 가방 추가하기 페이지 리팩토링
