+++ src/app/mybag/add/page.tsx
@@ -1,19 +1,16 @@
 'use client';
 
-import {
-    getScheduleTravelBagList,
-    getTravelBagList,
-    getTravelPlanList,
-    makeNewTravelBag
-} from '@/apis/bag';
+import React, { useState } from 'react';
+
+import { useRecoilValue } from 'recoil';
+
+import { makeNewTravelBag } from '@/apis/bag';
 import OtherSchedule from '@/components/detailschedule/OtherSchedule';
 import NewBagModal from '@/components/modal/NewBagModal';
 import BagList from '@/components/mybag/BagList';
 import EmptyBag from '@/components/mybag/EmptyBag';
 import { planIDState } from '@/states/schedule';
 import { IBag } from '@/types/bag';
-import React, { useEffect, useRef, useState } from 'react';
-import { useRecoilValue } from 'recoil';
 
 const NewBag = () => {
     const [bagList, setBagList] = useState<IBag[]>([]);
@@ -34,31 +31,25 @@ const NewBag = () => {
         setIsNewBagModal(false);
     };
 
-    useEffect(() => {
-        if (planID !== -1) {
-            getScheduleTravelBagList(planID).then((data) => {
-                setBagList(data);
-            });
-        }
-    }, [planID]);
+    const handleOpenModal = () => setIsNewBagModal(true);
 
     return (
         <>
-            <div className='flex items-end h-36 gap-4 text-grey mb-4'></div>
+            <div className='h-36 gap-4 text-grey mb-4'></div>
             <OtherSchedule href='mybag' register={false} top='top-[340px]' />
-            <div className='h-96'>
-                <div className='flex h-full'>
-                    <BagList bagList={bagList} />
-                    <EmptyBag setIsNewBagModal={setIsNewBagModal} />
-
-                    {isNewBagModal && (
-                        <NewBagModal
-                            setIsModal={setIsNewBagModal}
-                            handleAddNewBag={handleAddNewBag}
-                        />
-                    )}
-                </div>
+            <div className='flex h-full py-12'>
+                {bagList.length > 0 &&
+                    bagList.map((bag) => (
+                        <BagList key={bag.bag_name} bag={bag} />
+                    ))}
+                <EmptyBag handleOpenModal={handleOpenModal} />
             </div>
+            {isNewBagModal && (
+                <NewBagModal
+                    setIsModal={setIsNewBagModal}
+                    handleAddNewBag={handleAddNewBag}
+                />
+            )}
         </>
     );
 };
