+++ src/components/mybag/EmptyBag.tsx
@@ -1,25 +1,30 @@
-import Image from 'next/image';
 import React from 'react';
+
+import Image from 'next/image';
+
 import RoundBtn from '../layout/roundBtn';
 
-const EmptyBag = ({ setIsNewBagModal }: any) => {
+interface EmptyBagProps {
+    handleOpenModal: () => void;
+}
+
+const EmptyBag = ({ handleOpenModal }: EmptyBagProps) => {
     return (
-        <>
-            <div
-                className='flex justify-center items-center basis-[25%]'
-                onClick={() => setIsNewBagModal(true)}
-            >
-                <div className='flex flex-col items-center gap-8 cursor-pointer'>
-                    <Image
-                        src='/images/bag.svg'
-                        alt='my-bag'
-                        width={150}
-                        height={150}
-                    />
-                    <RoundBtn label='가방 추가하기' color='bg-lightgrey' />
-                </div>
-            </div>
-        </>
+        <div className='flex flex-col justify-center items-center gap-8 px-20'>
+            <button onClick={handleOpenModal}>
+                <Image
+                    src='/images/bag.svg'
+                    alt='my-bag'
+                    width={150}
+                    height={150}
+                />
+            </button>
+            <RoundBtn
+                label='가방 추가하기'
+                color='bg-lightgrey'
+                onClick={handleOpenModal}
+            />
+        </div>
     );
 };
